Move some stuff to swapchain.rs
